requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Time series analysis & forecasting
statsmodels>=0.14.0
pmdarima>=2.0.0  # Auto ARIMA

# Association mining
mlxtend>=0.22.0  # For Apriori, FP-Growth

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Interface
streamlit>=1.28.0

# Clustering & anomaly detection
scipy>=1.10.0
hdbscan>=0.8.0  # Optional: better clustering

# Notebook execution
jupyter>=1.0.0
ipykernel>=6.0.0
papermill>=2.4.0  # For automated notebook execution

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0  # Progress bars
python-dateutil>=2.8.0

# Optional: GUI/Demo
# streamlit>=1.25.0
# gradio>=3.35.0
